17h15 15/08

+++ module2/src/java_collection_framework/bai_tap/quan_li_san_pham/SanPham.java
@@ -2,12 +2,12 @@ package java_collection_framework.bai_tap.quan_li_san_pham;
 
 public class SanPham {
     private String tenSanPham;
-    private int giaSanPham;
+    private float giaSanPham;
 
     public SanPham() {
     }
 
-    public SanPham(String tenSanPham, int giaSanPham) {
+    public SanPham(String tenSanPham, float giaSanPham) {
         this.tenSanPham = tenSanPham;
         this.giaSanPham = giaSanPham;
     }
@@ -20,11 +20,11 @@ public class SanPham {
         this.tenSanPham = tenSanPham;
     }
 
-    public int getGiaSanPham() {
+    public float getGiaSanPham() {
         return giaSanPham;
     }
 
-    public void setGiaSanPham(int giaSanPham) {
+    public void setGiaSanPham(float giaSanPham) {
         this.giaSanPham = giaSanPham;
     }
